[refactor] re-organize file structure.
